pyproject.toml
[project]
name = "study-for-strands-agents"
version = "0.1.0"
description = "Strands Agents マルチエージェントパターンの習作"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
]

[project.optional-dependencies]
a2a = [
    "strands-agents[a2a]",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pyright>=1.1",
    "ruff>=0.8",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
reportMissingImports = "warning"
